feat: ToolsHub window
Added a hub window to show all the tool options and streamlined some proccesses.
Added a toggle for the auto prefixer.
Added a rename history for the auto prefixer with adjustable history length to show.
Removed popup display for singleton import, instead uses a new temp page on the tool hub.

-might be bad on performance in OnGUI calls >.<